factor out jsProject
