+++ package.json
@@ -54,6 +54,7 @@
     "ui/tree",
     "ui/@build/tsProject",
     "ui/@build/tsPlugins",
+    "ui/@build/jsProject",
     "ui/@types/lichess"
   ]
 }
