+++ ui/@build/jsProject/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@build/jsProject",
+  "version": "2.0.0",
+  "private": true,
+  "author": "Thibault Duplessis",
+  "license": "AGPL-3.0",
+  "main": "index.js",
+  "dependencies": {
+    "gulp": "^4",
+    "vinyl-source-stream": "^2",
+    "vinyl-buffer": "^1",
+    "ansi-colors": "^1",
+    "fancy-log": "^1",
+    "watchify": "^3",
+    "browserify": "^16",
+    "gulp-terser": "^1",
+    "gulp-size": "^3"
+  }
+}
