+++ ui/editor/gulpfile.js
@@ -1,3 +1,3 @@
-require('../gulp/jsProject.js')('LichessEditor', 'lichess.editor', __dirname);
+require('@build/jsProject')('LichessEditor', 'lichess.editor', __dirname);
 
 require('../gulp/cssProject.js')(__dirname);
